requirements.txt
# Lista zależności Python
kafka-python==2.0.2
faker==19.13.0
sentence-transformers==2.2.2
qdrant-client==1.6.5
psycopg2-binary==2.9.9
requests==2.31.0
rich==13.7.1
python-dotenv==1.0.1